+++ hydrodatasource/downloader/__init__.py
@@ -4,10 +4,10 @@ Date: 2023-10-06 20:50:41
 LastEditTime: 2023-10-14 19:30:14
 LastEditors: Wenyu Ouyang
 Description: Top-level package for wis-downloader
-FilePath: \hydrodata\hydrodata\downloader\__init__.py
+FilePath: \hydrodatasource\hydrodatasource\downloader\__init__.py
 Copyright (c) 2023-2024 Jeff Zhu. All rights reserved.
 """
 
-__author__ = """Jeff Zhu"""
-__email__ = "zjf014@gmail.com"
-__version__ = "0.0.1"
+__author__ = """Wenyu Ouyang"""
+__email__ = "wenyuouyang@outlook.com"
+__version__ = '0.0.1'