+++ tests/test_reader_basins.py
@@ -5,7 +5,7 @@ Date: 2023-10-25 15:16:21
 LastEditTime: 2024-03-28 08:40:25
 LastEditors: Wenyu Ouyang
 Description: Tests for reading basins' data
-FilePath: \hydrodata\tests\test_reader_basins.py
+FilePath: \hydrodatasource\tests\test_reader_basins.py
 Copyright (c) 2023-2024 Wenyu Ouyang. All rights reserved.
 """
 import os.path