+++ tests/test_rr_event_iden.py
@@ -4,7 +4,7 @@ Date: 2023-10-28 09:23:22
 LastEditTime: 2024-03-28 08:40:17
 LastEditors: Wenyu Ouyang
 Description: Test for rainfall-runoff event identification
-FilePath: \hydrodata\tests\test_rr_event_iden.py
+FilePath: \hydrodatasource\tests\test_rr_event_iden.py
 Copyright (c) 2023-2024 Wenyu Ouyang. All rights reserved.
 """